+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.7.0 (unveröffentlicht)
 - `RF-161`: Bibliotheken binden genutzte Bibliotheken direkt ein und nicht mehr über BOM-Bibliotheken
+- `IFS-66`: Im Status "Abbruch" kann ein neuer Lauf nur mit "Restart" erfolgen, auch wenn im abgebrochenen Lauf zuvor kein Satz verarbeitet wurde.
 
 # v1.6.0
 - `IFS-17`: Umbenennung der Artifact-ID und Group-ID